Create ResourceProcessor
